+++ app/config/theme/index.ts
